removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12313 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/output/MSVTKExport.cpp
@@ -87,7 +87,7 @@ MSVTKExport::write(OutputDevice& of, SUMOTime timestep) {
 		of.closeTag();
 
 		of.openTag("Verts")  << ">\n";
-		of.openTag("DataArray") << " type=\"Int64\" Name=\"connectivity\" format=\"ascii\" >" << getOffset(speed.size());
+		of.openTag("DataArray") << " type=\"Int64\" Name=\"connectivity\" format=\"ascii\" >" << getOffset((int) speed.size());
 		of.closeTag();
 
 		of.openTag("DataArray") << " type=\"Int64\" Name=\"offsets\" format=\"ascii\" >" << speed.size();
